+++ pennylane_lightning/core/src/simulators/lightning_qubit/measurements/MeasurementsLQubit.hpp
@@ -598,7 +598,7 @@ class Measurements final
         const std::size_t n_wires = wires.size();
         std::vector<std::size_t> samples(num_samples * n_wires);
         this->setRandomSeed();
-        discrete_random_variable<PrecisionT> drv{this->rng, probs(wires)};
+        DiscreteRandomVariable<PrecisionT> drv{this->rng, probs(wires)};
         for (std::size_t s = 0; s < num_samples; s++) {
             const std::size_t idx = drv();
             for (std::size_t j = 0; j < n_wires; j++) {